Only update chisel tools when we update runtime dockerfiles
