+++ eng/update-dependencies/UpdateDependencies.cs
@@ -418,8 +418,11 @@ namespace Dotnet.Docker
                 new BaseUrlUpdater(RepoRoot, Options),
                 new MinGitUrlUpdater(RepoRoot, minGitRelease),
                 new MinGitShaUpdater(RepoRoot, minGitRelease),
-                new BasicVariableUpdater(RepoRoot, "runtime", "chisel|ref", chiselRef),
-                new BasicVariableUpdater(RepoRoot, "runtime", "rocks-toolbox|ref", rocksToolboxRef)
+                // Chisel updaters must be listed before runtime version
+                // updaters because they check the manifest for whether the
+                // runtime versions are being updated or not
+                new ChiselToolUpdater(RepoRoot, "chisel|6.0|8.0|ref", Options.DockerfileVersion, chiselRef),
+                new ChiselToolUpdater(RepoRoot, "rocks-toolbox|6.0|8.0|ref", Options.DockerfileVersion, rocksToolboxRef)
             };
 
             foreach (string productName in Options.ProductVersions.Keys)
